requirements.txt
absl-py==2.0.0
antialiased-cnns==0.3
asttokens==2.4.0
backcall==0.2.0
blinker==1.6.2
brotlipy==0.7.0
cachetools==5.3.1
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Flask==2.3.3
fonttools==4.42.1
google-auth==2.23.0
google-auth-oauthlib==0.4.6
grpcio==1.58.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipycanvas==0.13.1
ipyevents==2.0.2
ipython==8.12.2
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==7.4.9
jupyter_core==5.3.1
jupyterlab-widgets==3.0.9
kaolin @ file:///home/ljw/Downloads/kaolin-0.14.0-cp38-cp38-linux_x86_64.whl#sha256=f827db327e749c9c5ae4ab50ec1fed4302b3c89816a18344e2bf7a427194a63a
kiwisolver==1.4.5
lpips==0.1.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.5.3.56
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///croot/pillow_1695134008276/work
platformdirs==3.10.0
plyfile==0.7.4
prompt-toolkit==3.0.39
protobuf==3.20.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.16.1
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
PyWavelets==1.4.1
pyzmq==24.0.1
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.2
scipy==1.7.1
six==1.16.0
smplx==0.1.28
stack-data==0.6.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tifffile==2023.7.10
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
trimesh==3.10.7
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
urllib3 @ file:///croot/urllib3_1686163155763/work
usd-core==23.5
wcwidth==0.2.6
Werkzeug==2.3.7
widgetsnbextension==4.0.9
zipp==3.17.0
